wrap eventlist to mainpage
